Make the data durable on disk.
This commit makes pg_reset_page_lsn sync the target directory and
all the contents in it to the disk by default after resetting the LSN of
the target relation pages.

This commit also allows pg_reset_page_lsn to accept -N option.
If this is specified, pg_reset_page_lsn returns without waiting for
all files to be written safely to disk.